infinite scroll frontend

+++ client/src/layouts/tables/index.js
@@ -35,6 +35,7 @@ import DashboardNavbar from "examples/Navbars/DashboardNavbar";
 import Footer from "examples/Footer";
 import DataTable from "examples/Tables/DataTable";
 import CreateAuthorsDialog from "./CreateAuthorDialog"
+import CreateTutorialsDialog from "./CreateTutorialDialog"
 
 // Data
 import authorsTableData from "layouts/tables/data/authorsTableData";
@@ -42,8 +43,9 @@ import projectsTableData from "layouts/tables/data/projectsTableData";
 
 import { useCallback, useEffect,useState } from "react";
 import { deleteAuthor } from "./data/deleteAuthor";
+import { deleteTutorial } from "./data/deleteTutorial";
 
-const agGridOptions = {
+const agGridAuthorsOptions = {
   columnDefs: [
     { headerName: 'id', field: 'id' },
     { headerName: 'name', field: 'name' },
@@ -51,37 +53,103 @@ const agGridOptions = {
     { headerName: 'starsCount', field: 'starsCount' },
     { headerName: 'birthDate', field: 'birthDate' }
   ],
+  rowBuffer: 0,
   rowSelection: 'single',
-  onSelectionChanged: onSelectionChanged
-
+  onSelectionChanged: onAuthorsSelectionChanged,
+  rowModelType: 'infinite',
+  // how big each page in our page cache will be, default is 100
+  cacheBlockSize: 5, //100,
+  // how many extra blank rows to display to the user at the end of the dataset,
+  // which sets the vertical scroll and then allows the grid to request viewing more rows of data.
+  // default is 1, ie show 1 row.
+  cacheOverflowSize: 2,
+  // how many server side requests to send at a time. if user is scrolling lots, then the requests
+  // are throttled down
+  maxConcurrentDatasourceRequests: 1,
+  // how many rows to initially show in the grid. having 1 shows a blank row, so it looks like
+  // the grid is loading from the users perspective (as we have a spinner in the first col)
+  infiniteInitialRowCount: 5, //100
+  // how many pages to store in cache. default is undefined, which allows an infinite sized cache,
+  // pages are never purged. this should be set for large data to stop your browser from getting
+  // full of data
+  maxBlocksInCache: 5//10
 
 }
-function onSelectionChanged() {
-  const selectedRows = agGridOptions.api.getSelectedRows();
 
-  console.log(selectedRows)
+
+
+function onAuthorsSelectionChanged() {
+  const selectedRows = agGridAuthorsOptions.api.getSelectedRows();
 }
+
 function Tables() {
   const { columns, rows } = authorsTableData();
   const { columns: pColumns, rows: pRows } = projectsTableData();
 
-  
-
-  const initTable = useCallback(async () => {
+  const initAuthorsTable = useCallback(async () => {
     const authors = await loadAuthors();
-    // const tutorials = await loadTutorials();
-    var eGridDiv = document.querySelector('#myGrid');
-    agGridOptions.rowData = authors
-   new AgGrid(eGridDiv, agGridOptions);
-   
-   
+    const gridDiv = document.querySelector('#authorsGrid');
+    //agGridAuthorsOptions.rowData = authors
+    const dataSource = {
+      rowCount: undefined, // behave as infinite scroll
+
+      getRows: (params) => {
+        console.log('asking for ' + params.startRow + ' to ' + params.endRow);
+
+        // At this point in your code, you would call the server.
+        // To make the demo look real, wait for 500ms before returning
+        setTimeout(function () {
+          // take a slice of the total rows
+          const rowsThisPage = authors.slice(params.startRow, params.endRow);
+          // if on or after the last page, work out the last row.
+          let lastRow = -1;
+          if (authors.length <= params.endRow) {
+            lastRow = authors.length;
+          }
+          // call the success callback
+          params.successCallback(rowsThisPage, lastRow);
+        }, 500);
+      },
+    };
+
+    new AgGrid(gridDiv, agGridAuthorsOptions); 
+
+    agGridAuthorsOptions.api.setDatasource(dataSource);
+
+
+
+}, []);
+
+const agGridTutorialsOptions = {
+  columnDefs: [
+    { headerName: 'price', field: 'price' },
+    { headerName: 'pageCount', field: 'pageCount' },
+    { headerName: 'title', field: 'title' },
+    { headerName: 'description', field: 'description' },
+    { headerName: 'publishedDate', field: 'publishedDate' }
+  ],
+  rowSelection: 'single',
+  onSelectionChanged: onAuthorsSelectionChanged
+}
+
+function onTutorialsSelectionChanged() {
+  const selectedRows = agGridTutorialsOptions.api.getSelectedRows();
+}
+
+const initTutorialsTable = useCallback(async () => {
+  
+  const tutorials = await loadTutorials();
+  const eGridDiv = document.querySelector('#tutorialsGrid');
+  agGridTutorialsOptions.rowData = tutorials
+ new AgGrid(eGridDiv, agGridTutorialsOptions);
+ 
+ 
 }, []);
 
   useEffect(() => {
-    initTable();
-    
+    initAuthorsTable();
+    initTutorialsTable();
   }, []);
-
 
   const [isCreateAuthorDialogOpen, setIsCreateAuthorDialogOpen] = useState(false);
 
@@ -89,34 +157,61 @@ function Tables() {
     setIsCreateAuthorDialogOpen(true);
   };
 
-  const handleDeleteAuthorClick = () => {
-    const selectedRows = agGridOptions.api.getSelectedRows();
-    const selectedAuthor = selectedRows[0]
-    const confirmed = window.confirm(`Delete ${selectedAuthor.name}?`)
-    if(confirmed){
-      deleteAuthor(selectedAuthor.id)
-    }
+  const [isCreateTutorialDialogOpen, setIsCreateTutorialDialogOpen] = useState(false);
+
+  const handleAddTutorialClick = () => {
+    setIsCreateTutorialDialogOpen(true);
   };
 
   
 
+  const handleDeleteAuthorClick = async() => {
+
+    const selectedRows = agGridAuthorsOptions.api.getSelectedRows();
+
+    if (selectedRows.length === 0){
+      window.alert("Please select row")
+      return;
+    } 
+
+    const selectedAuthor = selectedRows[0]
+    const confirmed = window.confirm(`Delete ${selectedAuthor.name}?`)
+    if(confirmed){
+      await deleteAuthor(selectedAuthor.id)
+      window.location.reload()
+    }
+      
+
+  };
+
+  const handleDeleteTutorialClick = () => {
+    const selectedRows = agGridTutorialsOptions.api.getSelectedRows();
+    const selectedTutorial = selectedRows[0]
+    const confirmed = window.confirm(`Delete ${selectedTutorial.name}?`)
+    if(confirmed){
+      deleteAuthor(selectedTutorial.id)
+    }
+
+  };
+
+
   return (
     <DashboardLayout>
       <DashboardNavbar />
-        <MDBox pt={1} pb={3}>
-          <div display="flex" justify-content="space-between">      
-          <MDButton variant="gradient" color="dark" onClick={handleAddAuthorClick}>
-          <Icon sx={{ fontWeight: "bold" }}>add</Icon>
-          &nbsp;add new author
-          </MDButton>
+          <MDBox pt={1} pb={3}>
+            <div display="flex" justify-content="space-between">      
+            <MDButton variant="gradient" color="info" onClick={handleAddAuthorClick}>
+              <Icon sx={{ fontWeight: "bold" }}>add</Icon>
+              &nbsp;add new author
+            </MDButton>
 
-          <MDButton variant="gradient" color="dark">
-          <Icon sx={{ fontWeight: "bold" }}>add</Icon>
+          <MDButton variant="gradient" color="warning">
+          <Icon sx={{ fontWeight: "bold" }}>update</Icon>
           &nbsp;change new author
           </MDButton>
 
           <MDButton variant="gradient" color="dark" onClick={handleDeleteAuthorClick}>
-          <Icon sx={{ fontWeight: "bold" }}>add</Icon>
+          <Icon sx={{ fontWeight: "bold" }}>delete</Icon>
           &nbsp;delete author
           </MDButton>
           </div>
@@ -139,16 +234,25 @@ function Tables() {
               >
                 <MDTypography variant="h6" color="white">
                 Authors
-                <div id="myGrid" style={{height: "400px", width: "100%", borderRadius: "10px", overflow: "hidden"}} className="ag-theme-alpine"></div>
+                <div id="authorsGrid" style={{height: "200px", width: "100%", borderRadius: "10px", overflow: "hidden"}} className="ag-theme-alpine"></div>
                 </MDTypography>
               </MDBox>
             </Card>
           </Grid>
 
+         
           <Grid item xs={12}>
             <Card>
               <MDBox pt={3}>
-                  
+              <MDButton variant="gradient" color="dark" onClick={handleAddTutorialClick}>
+                <Icon sx={{ fontWeight: "bold" }}>add</Icon>
+                &nbsp;add new tutorial
+              </MDButton>
+              <MDButton variant="gradient" color="dark" onClick={handleDeleteTutorialClick}>
+                <Icon sx={{ fontWeight: "bold" }}>add</Icon>
+                &nbsp;delete tutorial
+              </MDButton>
+              <CreateTutorialsDialog isCreateTutorialDialogOpen={isCreateTutorialDialogOpen} setIsCreateTutorialDialogOpen={setIsCreateTutorialDialogOpen} />
               </MDBox>
               <MDBox
                 mx={2}
@@ -162,7 +266,7 @@ function Tables() {
               >
                 <MDTypography variant="h6" color="white">
                   Tutorials
-                <div id="myGrid" style={{height: "400px", width: "100%", borderRadius: "10px", overflow: "hidden"}} className="ag-theme-alpine"></div>
+                <div id="tutorialsGrid" style={{height: "400px", width: "100%", borderRadius: "10px", overflow: "hidden"}} className="ag-theme-alpine"></div>
                 </MDTypography>
               </MDBox>
             </Card>
